+++ webkit/media/webmediaplayer_ms.h
@@ -75,19 +75,19 @@ class WebMediaPlayerMS
   virtual bool supportsFullscreen() const OVERRIDE;
   virtual bool supportsSave() const OVERRIDE;
   virtual void seekFloat(float seconds);
-  virtual void seek(float seconds);
+  virtual void seek(double seconds);
   virtual void setEndTimeFloat(float seconds);
-  virtual void setEndTime(float seconds);
+  virtual void setEndTime(double seconds);
   virtual void setRateFloat(float rate);
-  virtual void setRate(float rate);
+  virtual void setRate(double rate);
   virtual void setVolumeFloat(float volume);
-  virtual void setVolume(float volume);
+  virtual void setVolume(double volume);
   virtual void setVisible(bool visible) OVERRIDE;
   virtual void setPreload(WebKit::WebMediaPlayer::Preload preload) OVERRIDE;
   virtual bool totalBytesKnown() OVERRIDE;
   virtual const WebKit::WebTimeRanges& buffered() OVERRIDE;
   virtual float maxTimeSeekableFloat() const;
-  virtual float maxTimeSeekable() const;
+  virtual double maxTimeSeekable() const;
 
   // Methods for painting.
   virtual void setSize(const WebKit::WebSize& size) OVERRIDE;
@@ -107,9 +107,9 @@ class WebMediaPlayerMS
   virtual bool paused() const OVERRIDE;
   virtual bool seeking() const OVERRIDE;
   virtual float durationFloat() const;
-  virtual float duration() const;
+  virtual double duration() const;
   virtual float currentTimeFloat() const;
-  virtual float currentTime() const;
+  virtual double currentTime() const;
 
   // Get rate of loading the resource.
   virtual int32 dataRate() const OVERRIDE;
@@ -126,7 +126,7 @@ class WebMediaPlayerMS
   virtual WebKit::WebMediaPlayer::MovieLoadType movieLoadType() const OVERRIDE;
 
   virtual float mediaTimeForTimeValueFloat(float timeValue) const;
-  virtual float mediaTimeForTimeValue(float timeValue) const;
+  virtual double mediaTimeForTimeValue(double timeValue) const;
 
   virtual unsigned decodedFrameCount() const OVERRIDE;
   virtual unsigned droppedFrameCount() const OVERRIDE;
